+++ src/ReinforcementLearningExperiments/src/experiments/MARL/DQN_mpe_simple.jl
@@ -7,14 +7,14 @@
 # ---
 
 using PyCall
-using ReinforcementLearning
+using ReinforcementLearningCore, ReinforcementLearningBase, ReinforcementLearningZoo
 using Flux
 using Flux: glorot_uniform
 
 using StableRNGs: StableRNG
 using Flux.Losses: huber_loss
 
-function RL.Experiment(
+function RLCore.Experiment(
     ::Val{:JuliaRL},
     ::Val{:DQN},
     ::Val{:MPESimple};
@@ -39,7 +39,7 @@ function RL.Experiment(
                         );
                         sync_freq=100
                     ),
-                    optimiser=ADAM(),
+                    optimiser=Adam(),
                 ),
                 n=n,
                 γ=γ,